Context necessary so as to not make many api calls'
