+++ P1/entertainment-app/src/hooks/useBookmarks.js
@@ -1,24 +1,7 @@
-import { useUser } from "@clerk/clerk-react";
-import { useQuery } from "@tanstack/react-query";
-
-const backendRootUrl = import.meta.env.VITE_BACKEND_URL
+import { BookmarkContext } from '../components/BookmarkProvider.jsx';
+import { useContext } from 'react';
 
 export const useBookmarks = () => {
-    const { user } = useUser();
-    let userEmail = user.primaryEmailAddress.emailAddress
-
-    const {isStale, data: bookmarks} = useQuery({
-        queryKey: [`bookmarks`],
-        queryFn: async () => {
-            const res = await fetch(backendRootUrl + "api/" + `users/bookmarks/?email=${userEmail}`);
-            const data = await res.json();
-            let bookmarkIds = {};
-            data.forEach(bookmark => {
-                bookmarkIds[bookmark.id] = true;
-            });
-            return bookmarkIds;
-        },
-    });
-
-    return {bookmarks: bookmarks, isStale: isStale};
+    const bookmarks = useContext(BookmarkContext);
+    return { bookmarks };
 }
